Add function to check if a number is duplicated

+++ push_swap/includes/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: alperrot <alperrot@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/19 09:04:02 by alperrot          #+#    #+#             */
-/*   Updated: 2024/05/20 12:05:12 by alperrot         ###   ########.fr       */
+/*   Updated: 2024/05/20 12:21:17 by alperrot         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int		check_args(char **av);
 t_stack	*create_node(long nb);
 void	five_sort(t_stack **s_a, t_stack **s_b);
 void	free_stack(t_stack *s);
-int		ft_atoi(const char *str);
+long	ft_atoi(const char *str);
 char	*ft_joinchar(char *s1, char const s2);
 int		get_max(t_stack *s, int max);
 int		has_space(char *str);
